# docs/openapi.yaml
openapi: 3.1.0
info:
  title: BrainVault API
  version: 1.0.0
  description: Block-based note app server API.
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PublicUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
          description: Login ID. This is not an email address.
        name:
          type: [string, 'null']
        avatarData:
          type: [string, 'null']
          description: PNG, JPEG, or WebP data URL used as the profile image.
        preferredLanguage:
          type: [string, 'null']
          enum: [en, ja, ko, fr, de, es, pt, null]
        defaultCollectionIcon:
          type: [string, 'null']
          maxLength: 32
          description: Emoji used for the built-in default collection.
        createdAt:
          type: string
        updatedAt:
          type: string
...
    SharePageInput:
      type: object
      required: [username]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 40
          pattern: '^[a-zA-Z0-9._-]+$'
          description: Existing BrainVault login ID. The owner cannot share a page with itself.
    PageShare:
      type: object
      required: [user, permission, sharedAt]
      properties:
        user:
          $ref: '#/components/schemas/PublicUser'
        permission:
          type: string
          const: EDIT
        sharedAt:
          type: string
          format: date-time
    CollaborationBlock:
      type: object
      required: [id, type, markdown, checked, parentBlockId, sortOrder]
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 64
        type:
          type: string
          enum: [MARKDOWN, HEADING_1, HEADING_2, HEADING_3, TODO, QUOTE, CALLOUT, TABLE, KANBAN, DATABASE, BOOKMARK, AI_CHAT, MATH, CODE, DIVIDER, IMAGE, ATTACHMENT]
        markdown:
          type: string
          maxLength: 20000
        checked:
          type: boolean
        parentBlockId:
          type: [string, 'null']
          maxLength: 64
        sortOrder:
          type: integer
          minimum: 0
          maximum: 2147483647
        metadata:
          type: [object, 'null']
          additionalProperties: true
    CollaborationSnapshotInput:
      type: object
      required: [title, blocks, updateId]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 160
          description: Must contain at least one non-whitespace character.
        blocks:
          type: array
          maxItems: 10000
          items:
            $ref: '#/components/schemas/CollaborationBlock'
        deletedAttachmentIds:
          type: array
          maxItems: 10000
          default: []
          items:
            type: string
            minLength: 1
            maxLength: 64
        updateId:
          type: integer
          minimum: 1
          description: Exact latest persisted Yjs update ID applied by the client.
    CollaborationSession:
      type: object
      required: [pageId, protocol, ticketProtocol, path, expiresIn, access, collaboration, document]
      properties:
        pageId:
          type: string
        protocol:
          type: string
          const: brainvault-yjs-v1
        ticketProtocol:
          type: string
          description: Short-lived page-scoped credential in the form brainvault-ticket.<JWT>. Send it as a WebSocket subprotocol, never as a query parameter.
        path:
          type: string
          example: /api/collaboration/page_123
        expiresIn:
          type: integer
          description: Ticket lifetime in seconds.
        access:
          type: object
          additionalProperties: true
        collaboration:
          type: object
          additionalProperties: true
        document:
          type: object
          required: [title, blocks]
          properties:
            title:
              type: string
            blocks:
              type: array
              items:
                type: object
                additionalProperties: true
paths:
...
  /api/pages/{pageId}/shares:
    get:
      summary: List users invited to edit a page
      description: Owner only. Collections and archived pages cannot be shared.
      parameters:
        - in: path
          name: pageId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Current editor grants
          content:
            application/json:
              schema:
                type: object
                required: [shares, count]
                properties:
                  shares:
                    type: array
                    items:
                      $ref: '#/components/schemas/PageShare'
                  count:
                    type: integer
                    minimum: 0
        '404':
          description: Page not owned by the authenticated user
    post:
      summary: Invite an existing BrainVault user as a page editor
      description: Owner only. Creating the first grant initializes a fresh collaboration history from the current canonical page snapshot.
      parameters:
        - in: path
          name: pageId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SharePageInput'
      responses:
        '201':
          description: Editor grant created
          content:
            application/json:
              schema:
                type: object
                required: [share, count]
                properties:
                  share:
                    $ref: '#/components/schemas/PageShare'
                  count:
                    type: integer
                    minimum: 1
        '400':
          description: Invalid account, self-share, collection, or archived page
        '409':
          description: The user already has access
  /api/pages/{pageId}/shares/{userId}:
    delete:
      summary: Remove a page editor
      description: Owner only. Active sockets for the removed user are closed immediately. Removing the final grant requires every accepted Yjs update to have been materialized.
      parameters:
        - in: path
          name: pageId
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Grant removed
          content:
            application/json:
              schema:
                type: object
                required: [removed, count]
                properties:
                  removed:
                    type: boolean
                    const: true
                  count:
                    type: integer
                    minimum: 0
        '409':
          description: Accepted collaboration updates are still pending materialization
  /api/pages/{pageId}/collaboration/session:
    post:
      summary: Create a short-lived authenticated collaboration session
      description: Available to the owner and invited editors while the page has at least one editor grant. The response contains the canonical relational snapshot and the required WebSocket subprotocols for WS /api/collaboration/{pageId}.
      parameters:
        - in: path
          name: pageId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Collaboration ticket and canonical document
          headers:
            Cache-Control:
              schema:
                type: string
              description: private, no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollaborationSession'
        '403':
          description: Authenticated user has no access to this page
        '409':
          description: Collaboration is disabled, archived, or unsupported for a collection
  /api/pages/{pageId}/collaboration/snapshot:
    put:
      summary: Materialize a consistent Yjs document into the relational page and block tables
      description: Available to the owner and invited editors. The server locks the page, requires updateId to equal the latest persisted update, validates hierarchy and attachment provenance, and applies the snapshot atomically.
      parameters:
        - in: path
          name: pageId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollaborationSnapshotInput'
      responses:
        '200':
          description: Materialization state and canonical blocks
          content:
            application/json:
              schema:
                type: object
                required: [applied, materializedUpdateId, pageVersion, pageContentVersion, pageUpdatedAt, blocks]
                properties:
                  applied:
                    type: boolean
                  materializedUpdateId:
                    type: integer
                    minimum: 1
                  pageVersion:
                    type: integer
                    minimum: 1
                  pageContentVersion:
                    type: integer
                    minimum: 1
                  pageUpdatedAt:
                    type: string
                    format: date-time
                  blocks:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
        '400':
          description: Invalid title, block hierarchy, attachment operation, or document size
        '409':
          description: Snapshot update ID is stale, a block ID conflicts globally, or collaboration is disabled
...